pyproject.toml
[project]
name = "image-caption-generation-vit-from-scratch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "datasets>=4.5.0",
    "huggingface-hub>=0.36.2",
    "ipykernel>=6.31.0",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.57.6",
]
